adding a panel for every shape
in the previous commit there was too much repeating code so we decided to use a single class and pass different arguments to configure each CreatePanel.
If one needs to be different, we can always create another derived Panel class and pass it as a pointer that needs to be SetActive when the button is clicked.